fix: extract DS provider and wrap it around app
