+++ packages/design-system/src/lib/provider.tsx
@@ -1,0 +1,11 @@
+import { StyledEngineProvider, ThemeProvider } from '@mui/material'
+
+import { theme } from './muiTheme'
+
+export const Provider = ({ children }: { children: React.ReactNode }) => {
+  return (
+    <StyledEngineProvider injectFirst>
+      <ThemeProvider theme={theme}>{children}</ThemeProvider>
+    </StyledEngineProvider>
+  )
+}
